+++ packages/shared/locales/en/tutors.json
@@ -437,6 +437,7 @@
       "previous": "Previous",
       "next": "Next"
     },
-    "close": "Close"
+    "close": "Close",
+    "youMightAlsoLike": "You might also like"
   }
 }
