+++ packages/shared/locales/vi/tutors.json
@@ -437,6 +437,7 @@
       "previous": "Trước",
       "next": "Sau"
     },
-    "close": "Đóng"
+    "close": "Đóng",
+    "youMightAlsoLike": "Bạn có thể cũng thích"
   }
 }
